feat(api/v1/chat/*): added gemini 2.0 endpoint with 60s streaming window
