+++ .changeset/thick-kids-judge.md
@@ -1,0 +1,5 @@
+---
+"mantine-template": minor
+---
+
+Changed Vercel custom domain to proto.buckeye.chat for simplicity. Added Google genai (gemini 2.0 flash) endpoint with max 60s function duration to ensure full stream.
