+++ .env.example
@@ -1,0 +1,3 @@
+GOOGLE_GENAI_API_KEY=
+
+OPENAI_API_KEY=
