+++ package.json
@@ -49,13 +49,21 @@
     "start": "next start"
   },
   "dependencies": {
+    "@agentic/duck-duck-go": "^7.6.4",
+    "@agentic/firecrawl": "^7.6.4",
+    "@agentic/weather": "^7.6.4",
+    "@ai-sdk/google": "^1.2.10",
+    "@ai-sdk/openai": "^1.3.10",
+    "@ai-sdk/react": "^1.2.8",
     "@mantine/core": "^7.17.4",
     "@mantine/hooks": "^7.17.4",
     "@mantine/notifications": "^7.17.4",
     "@tabler/icons-react": "^3.31.0",
+    "ai": "^4.3.5",
     "next": "^15.3.0",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "zod": "^3.24.2"
   },
   "devDependencies": {
     "@biomejs/biome": "^1.9.4",
@@ -93,6 +101,7 @@
       "@biomejs/biome",
       "core-js-pure",
       "esbuild",
+      "openai-zod-to-json-schema",
       "sharp"
     ]
   }
